refactor: 컴포넌트 폴더별 분리 refactor: 사용하지 않는 코드 삭제 refactor: alias로 경로 변경하여 코드 가독성 향상
