+++ src/components/Footer.vue
@@ -1,0 +1,14 @@
+<template>
+  <footer class="border-t-[1px] border-black dark:border-white">
+    <p>우리FISA 프론트엔드 기술 세미나</p>
+    <p>Copyright @ 2023 그린라이트 - All rights reserved.</p>
+  </footer>
+</template>
+<style scoped>
+footer {
+  padding: 30px;
+}
+p {
+  text-align: center;
+}
+</style>
