+++ src/views/OthersView.vue
@@ -1,25 +1,17 @@
 <script setup>
-import ContentTitle from '../components/ContentTitle.vue';
-import ContentBox from '../components/ContentBox.vue';
+import ContentTitle from '@/components/ContentTitle.vue';
+import ContentBox from '@/components/ContentBox.vue';
 </script>
 <template>
-    <ContentTitle title="Vanilla JavaScript 예시" />
-    <ContentBox>
+  <ContentTitle title="Vanilla JavaScript 예시" />
+  <ContentBox> </ContentBox>
 
-    </ContentBox>
+  <ContentTitle title="Next.js 예시" />
+  <ContentBox> </ContentBox>
 
-    <ContentTitle title="Next.js 예시" />
-    <ContentBox>
+  <ContentTitle title="소프트웨어 디자인 패턴" />
+  <ContentBox> </ContentBox>
 
-    </ContentBox>
-
-    <ContentTitle title="소프트웨어 디자인 패턴" />
-    <ContentBox>
-
-    </ContentBox>
-
-    <ContentTitle title="참고문서" />
-    <ContentBox>
-
-    </ContentBox>
+  <ContentTitle title="참고문서" />
+  <ContentBox> </ContentBox>
 </template>
